Rename review_step to review_logs_step
