+++ x-pack/plugins/upgrade_assistant/public/application/components/overview/_index.scss
@@ -1,2 +1,2 @@
-@import 'review_step/index';
+@import 'review_logs_step/index';
 @import 'fix_deprecation_logs_step/index';
