+++ x-pack/plugins/upgrade_assistant/public/application/components/overview/review_logs_step/_index.scss
