+++ x-pack/plugins/upgrade_assistant/public/application/components/overview/review_logs_step/es_stats/index.ts
